Update README.mdown
Hi Gaurav. I have added the content for variables, primitive types and operators. As usual it is incomplete because of the formatting. 
This time i am facing problem on adding the table. There are two tables in the document as of now. If possible can you do the formatting for one table. which I will take as reference and use it for whole document.